+++ app/src/main/java/com/example/administrator/counter/Entrance.java
@@ -1,15 +1,22 @@
 package com.example.administrator.counter;
 
 import android.app.Activity;
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ProgressBar;
 import android.widget.Toast;
+
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.GooglePlayServicesUtil;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -32,6 +39,10 @@ public class Entrance extends Activity {
     private EditText user_id;
     private EditText user_pw;
     private JSONObject jobj;
+
+    // GCM 관련 변수
+    private static final int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
+    private BroadcastReceiver mRegistrationBroadcastReceiver;
 
     ProgressBar pb1;
 
@@ -62,6 +73,7 @@ public class Entrance extends Activity {
             Log.d(TAG, "등록된 PW: " + prefs.getString("password", ""));
             user_id.setText(prefs.getString("id", ""));
             user_pw.setText(prefs.getString("password", ""));
+            String token = prefs.getString("gcm_token", "");
 
             ApplicationClass.app = (ApplicationClass)getApplicationContext();
             // 로그인 통신 실행
@@ -69,7 +81,10 @@ public class Entrance extends Activity {
                 String userID = user_id.getText().toString();
                 String userPW = user_pw.getText().toString();
 
-                jobj = new JSONObject("{ \"id\" : \"" + userID + "\",\"password\" : \"" + userPW + "\"}");
+                jobj = new JSONObject()
+                            .put("id", userID)
+                            .put("password", userPW)
+                            .put("reg_id", token);
 
                 // 통신 실행
                 new HttpHandler().loginUser(jobj.toString(), new MyCallback() {
@@ -91,28 +106,17 @@ public class Entrance extends Activity {
         login_btn.setOnClickListener(new View.OnClickListener() {
              @Override
              public void onClick(View v) {
-                 String userID = user_id.getText().toString();
-                 String userPW = user_pw.getText().toString();
-                 if (userID.equals("") || userPW.equals("")) {
-                     Toast toast = Toast.makeText(getApplicationContext(), "비밀번호 또는 아이디를 입력해 주세요.", Toast.LENGTH_LONG);
-                     toast.show();
+                 Log.d(TAG, "로그인 프로세스 시작");
+                 SharedPreferences prefs = getSharedPreferences("PrefName", MODE_PRIVATE);
+                 if (prefs.getString("gcm_token", "").isEmpty()) {
+                     Log.d(TAG, "GCM 토큰 생성");
+                     // GCM 토큰 생성
+                     getInstanceIdToken();
+                     // GCM 토큰생성 받는 리시버
+                     registBroadcastReceiver();
                  } else {
-                     ApplicationClass.app = (ApplicationClass)getApplicationContext();
-                     // 로그인 통신 실행
-                     try {
-                         jobj = new JSONObject("{ \"id\" : \"" + userID + "\",\"password\" : \"" + userPW + "\"}");
-
-                         // 통신 실행
-                         new HttpHandler().loginUser(jobj.toString(), new MyCallback() {
-                             @Override
-                             public void httpProcessing(JSONObject result) {
-                                 loginProcess(result);
-                             }
-                         });
-
-                     } catch (JSONException e) {
-                         e.printStackTrace();
-                     }
+                     Log.d(TAG, "GCM 토큰 생성안함");
+                     login();
                  }
              }
          });
@@ -129,6 +133,43 @@ public class Entrance extends Activity {
             }
         });
 
+    }
+
+    /**
+     * 로그인 프로세스 시작
+     */
+    public void login() {
+
+        SharedPreferences prefs = getSharedPreferences("PrefName", MODE_PRIVATE);
+        String token = prefs.getString("gcm_token", "");
+        String userID = user_id.getText().toString();
+        String userPW = user_pw.getText().toString();
+
+        if (userID.equals("") || userPW.equals("")) {
+            Toast toast = Toast.makeText(getApplicationContext(), "비밀번호 또는 아이디를 입력해 주세요.", Toast.LENGTH_LONG);
+            toast.show();
+        } else {
+            ApplicationClass.app = (ApplicationClass)getApplicationContext();
+
+            // 로그인 통신 실행
+            try {
+                jobj = new JSONObject()
+                        .put("id", userID)
+                        .put("password", userPW)
+                        .put("reg_id", token);
+
+                // 통신 실행
+                new HttpHandler().loginUser(jobj.toString(), new MyCallback() {
+                    @Override
+                    public void httpProcessing(JSONObject result) {
+                        loginProcess(result);
+                    }
+                });
+
+            } catch (JSONException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
 
@@ -176,5 +217,98 @@ public class Entrance extends Activity {
         } catch (JSONException e) {
             e.printStackTrace();
         }
+    }
+
+
+
+    // GCM 토큰 생성
+    /**
+     * Instance ID를 이용하여 디바이스 토큰을 가져오는 RegistrationIntentService를 실행한다.
+     */
+    public void getInstanceIdToken() {
+        if (checkPlayServices()) {
+            // Start IntentService to register this application with GCM.
+            Intent intent = new Intent(this, RegistrationIntentService.class);
+            startService(intent);
+            Log.d(TAG, "토큰 생성 서비스 실행");
+        }
+    }
+
+    /**
+     * LocalBroadcast 리시버를 정의한다. 토큰을 획득하기 위한 READY, GENERATING, COMPLETE 액션에 따라 UI에 변화를 준다.
+     */
+    public void registBroadcastReceiver(){
+        mRegistrationBroadcastReceiver = new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, Intent intent) {
+                String action = intent.getAction();
+
+
+                if(action.equals(QuickstartPreferences.REGISTRATION_READY)){
+                    // 액션이 READY일 경우
+                    // TODO
+                    Log.d(TAG, "READY");
+                } else if(action.equals(QuickstartPreferences.REGISTRATION_GENERATING)){
+                    // 액션이 GENERATING일 경우
+                    // TODO
+                    Log.d(TAG, "GENERATING");
+                } else if(action.equals(QuickstartPreferences.REGISTRATION_COMPLETE)){
+                    // 액션이 COMPLETE일 경우
+                    Log.d(TAG, "COMPLETE");
+
+                    SharedPreferences prefs = getSharedPreferences("PrefName", MODE_PRIVATE);
+                    SharedPreferences.Editor editor = prefs.edit();
+                    String token = intent.getStringExtra("token");
+                    editor.putString(token, "");
+                    Log.d(TAG, "토큰생성완료: "+token);
+
+                    login();
+
+                }
+
+            }
+        };
+    }
+
+    /**
+     * 앱이 실행되어 화면에 나타날때 LocalBoardcastManager에 액션을 정의하여 등록한다.
+     */
+    @Override
+    protected void onResume() {
+        super.onResume();
+        LocalBroadcastManager.getInstance(this).registerReceiver(mRegistrationBroadcastReceiver,
+                new IntentFilter(QuickstartPreferences.REGISTRATION_READY));
+        LocalBroadcastManager.getInstance(this).registerReceiver(mRegistrationBroadcastReceiver,
+                new IntentFilter(QuickstartPreferences.REGISTRATION_GENERATING));
+        LocalBroadcastManager.getInstance(this).registerReceiver(mRegistrationBroadcastReceiver,
+                new IntentFilter(QuickstartPreferences.REGISTRATION_COMPLETE));
+
+    }
+
+    /**
+     * 앱이 화면에서 사라지면 등록된 LocalBoardcast를 모두 삭제한다.
+     */
+    @Override
+    protected void onPause() {
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(mRegistrationBroadcastReceiver);
+        super.onPause();
+    }
+
+    /**
+     * Google Play Service를 사용할 수 있는 환경이지를 체크한다.
+     */
+    private boolean checkPlayServices() {
+        int resultCode = GooglePlayServicesUtil.isGooglePlayServicesAvailable(this);
+        if (resultCode != ConnectionResult.SUCCESS) {
+            if (GooglePlayServicesUtil.isUserRecoverableError(resultCode)) {
+                GooglePlayServicesUtil.getErrorDialog(resultCode, this,
+                        PLAY_SERVICES_RESOLUTION_REQUEST).show();
+            } else {
+                Log.i(TAG, "This device is not supported.");
+                finish();
+            }
+            return false;
+        }
+        return true;
     }
 }
